Updated openstack/openstack Project: openstack-infra/project-config  f5f964521d14f566ad9fd8e048b0b7bbdb9ae662
Fix a cyclic CRD problem

This patch set moves all common libraries and toolkit from OpenStack-Helm
to OpenStack-Helm-Infra, and updates the zuul-cloner call. This allows OS-H
to zuul-clone from a centralized repo, OS-H-Infra.  Previously, OS-H-Infra
can have dependencies on OS-H for the helm-toolkit, but OS-H is also
dependent on the OS-H-Infra for the infrastructure-related charts --
causing a potential cyclical dependencies.

Change-Id: Ic61cdc8391bfd3b5b205a0bde354eef7407c6f48