Updated openstack/openstack Project: openstack/openstack-manuals  a77ec96fcbdce8fb1f7d1f022b616efad54f844c
add a page listing all configuration reference guides

Add a new template page and replace the old link to the draft
config-reference with a link to that page.

Add support for a has_config_ref flag in the template generator.

Set has_config_ref for the service projects that are publishing to
/configuration now. Leave other projects commented out to make it easy
for projects to turn on their link when they are ready.

Change-Id: Ie9d6379688a91835af42a6cbf739d73fefcfdab4
Signed-off-by: Doug Hellmann <doug@doughellmann.com>