Updated openstack/openstack Project: openstack-infra/project-config  6fa62e938455d5c028726fae549e70eb482d6729
Copy zuul v3 inventory file to log dir

Debugging that the inventory is what we think it should be is really
helpful. We pass secrets in as vars files, so there should be no secret
data in this. It needs to be done in the base job though, since it needs
to be trusted to be able to copy the file.

Change-Id: Ic1239c4635bbda409ff033f0d5d2321964b14bf9