Updated openstack/openstack Project: openstack/keystone  3aa56823eb27e4e2bb88c2c70dcf2750a554d68b
Gear documentation towards a wider audience

Our documentation previously had a statement saying it was primarily
geared towards contributing developers. This commit removes that
statement and replaces it with a more general one. Subsequent patches
are going to be migrating a bunch of useful information for
contributors, developers, and operators alike.

Change-Id: I8cea91b755f53265c21ec72b01939a7c646d3756