Updated openstack/openstack Project: openstack/ironic  54d5335edddd4c75dd3c8b6dc344761cb95abd9e
Mock random generator for BackoffLoopingCall in IPMI unittests

Set power state methods now use BackoffLoopingCall to wait for the
desired power state. It uses random.SystemRandom gauss distribution
to determine the amount of time to wait for the next check. This
change adds a method to mock the result of the pseudo-random sleep
interval generation, so that the value to wait between power state
checks is 1.

Closes-Bug: 1702859
Change-Id: I9270a187fa0f413ba8a8a14f859b0fd65c439b6e