Updated openstack/openstack Project: openstack-infra/project-config  4cbb0ec31726b3c2e126223f639d02c4c78d7011
Create python-jobs-trusty job group

Only a few projects need trusty jobs, let's split up python-jobs so that
we only define jobs for those projects that need it - and not for
everybody.

This is the first part in the series, it moves python34 jobs over and
adds it for those projects that need it.

Change-Id: Ib268711bec56276a039db287dc42e8a02593b805