Updated openstack/openstack Project: openstack-infra/elastic-recheck  afd71d18263bd22bafb502942164b5727169af25
Remove query for fixed bug 1654393

This was a unit test leak of a global variable not
getting reset properly. That was fixed in Ocata but
something new is tripping the same failure, so let's
remove the query so people don't assume it's not their
patch that's causing a new problem.

Change-Id: Ifc7266c2f1caed9add609b68b9bb274a50592384
Related-Bug: #1654393