Updated openstack/openstack Project: openstack/python-cinderclient  eaee84097ab15338e8655b8bd1590a25993c09d5
UnboundLocalError on message-list

This bug for 'cinder message-list' has been fixed such that now
if the user gives wrong filters, then the command properly
prints WARNING message saying it is ignoring that filter. Previously,
for wrong filters which were not passed in the correct usage format,
the 'cinder message-list --filters' command was giving UnboundLocalError
exception.

The extract_filters() method was previously assuming that all the
message filters passed were in the format "<filter_name>=<value>",
because of which if the user passed something like:
	'cinder message-list --filters event_id:VOLUME_000002'
an UnboundLocalError was thrown.

The changes are made to the method such that if the user passes a
filter in invalid format, the cinder CLI ignores that filter while
providing the output.

Closes-Bug: #1696556

Change-Id: I545b3f28e545b380842ea003dc38ad70cb413aee