Updated openstack/openstack Project: openstack/horizon  3765a5827a4a1e371591393cf6d13558ceddd54b
Removed default value for 'Physical Network'

When creating a network there is currently a default value of 'default'
which causes the network creation to fail. It is nonsensical to have
a default value here as it will change depending on the provider
network type and needs to be set according to custom settings if the
user is not creating a simple local network which then does not require
the physical network value at all.

Change-Id: I448713d986f17ea1c00cd4dd22487291545cfe2a
Closes-Bug: #1417305