Updated openstack/openstack Project: openstack/cinder  544d13ef0a9397a18af506607150b0f2c2c3752c
Migrate the blockstorage admin-guide to Cinder

This patch is part of the docs migration for Cinder.
It is more or less a drag and drop of the docs from
openstack-manuals admin-guide directory.  I needed to
change some syntax to work with Cinder's more stringent
doc build.

Note that the purpose of this patch is just to get the
documentation back to an accessible location.  Later patches
will clean up the organization and content.

Change-Id: Ib3f9255e0f9f2ff42a0ee4126607ff319a3d901e