Updated openstack/openstack Project: openstack/tripleo-heat-templates  322f2dbd4c1d9c8617186bcb9211d7a1fe10ba3f
Copy only generated puppet files into the container

This solves a problem with bind-mounts when the containers are holding
files descriptors open.

At the same time this makes the template more robust to puppet changes
since new config files will be available in the containers without
needing to update the templates.

Partial-Bug: #1698323
Change-Id: Ia4ad6d77387e3dc354cd131c2f9756939fb8f736