Updated openstack/openstack Project: openstack/openstackdocstheme  2969174191a3180356fcefe9611b955aef606377
Add support for versionadded and deprecated

This patch adds support for the versionadded, versionchanged and
deprecated tags. It also reduces the rather aggressive margins and
padding around the admonitions, so that they stack a little better,
and adds a missing icon to the 'seealso' admonition.

Adds example admonitions to the Demo documentation.

Change-Id: Ieb064b26980f885549660a80e2a2578b71a5e7b7