Updated openstack/openstack Project: openstack/openstack-manuals  60c0a3ebbb90db5f6ed597270b5e4deafe8561be
add a new landing page to list all admin guides

Add an admin-guide.html page that lists all admin guides for all
projects that have them.

Replace the link to the central admin guide on the main page with a link
to the new templated page.

Add a flag to the project data schema to indicate when a project has an
admin guide.

Update the project data to set the has_admin_guide flag for projects
that already have them published.

Change-Id: I514fe5034c26a42cfb4225d521691c32c0ba9ef4
Signed-off-by: Doug Hellmann <doug@doughellmann.com>