Updated openstack/openstack Project: openstack/openstack-manuals  6e44dbd9630b5742bb0034cd4f6e8e9b7a064d97
add libraries to list of configuration reference guides

Add a column to show the libraries that provide common configuration
options to services.

Also add a link to oslo.config's general info to the top of the list of
links for configuring services, since that's where the shared content
about things like file format and reconfiguring running services moved
when we deleted the configuration reference guide.

Change-Id: I1dee42565aa57ec8b8ac10b5093f4956b0629d00
Signed-off-by: Doug Hellmann <doug@doughellmann.com>