Updated openstack/openstack Project: openstack/keystone  2c67130fe013556d4a1189616fd125dca2206644
Move caching docs into admin-guide

The admin-guide and configuration guide both had separate sections
dedicated to documenting caching. This patch attempts to consolidate
them into a single section in the admin-guide. It also takes the
opportunity to update a couple pieces of information that have
changed since writing the original docs in configuration.rst.

The old configuration section also reference a lot of configuration
options for caching. These were removed because they should be
documented in the configuration guide.

Change-Id: Ifb39075a6a78419e5c44c7e0871f3e1c055c2332