Updated openstack/openstack Project: openstack/governance  04c9047fe3e0a20a1c413fb3d49ff3a9dc5731fc
Remove remaining 'big tent' references

The words 'Big tent' have proven to be confusing to people.
Their initial meaning was that project teams can join OpenStack
if they are "one of us" (as defined by sharing our mission and
our principles). However, misinformed people (or people who
maliciously choose to be misinformed) took it to mean that
there is an absence of rules, and anything and anyone can call
themselves an OpenStack project.

While most references to the 'big tent' were replaced over time
(with references to the 'project structure reform of 2014', or
'official projects' depending on what the words were actually
used for), some mentions still remained. This commit cleans them
up.

“It's a beautiful thing, the destruction of words.”
― George Orwell, 1984

Change-Id: I249ee50e1651997d4248b32a6e30d0737192f119