Updated openstack/openstack Project: openstack/ironic  bc5efdf45973941347752d3d5eabbffcefad971f
Physical network aware VIF attachment

When attaching virtual interfaces to ironic ports and portgroups, we
need to take account of the physical network assigned to those ports and
portgroups. A neutron virtual interface has a set of physical networks
on which it can be attached which is governed by the segments of its
network (of which there may be more than one).

This change makes the ironic VIF attach API physical network-aware using
the physical network information added to the port object.

When selecting a port or portgroup to attach a virtual interface to, the
following ordered criteria are applied:

* Require ports or portgroups to have a physical network that is
  None or one of the VIF's allowed physical networks.
* Prefer ports or portgroups with a physical network field which
  is not None.
* Prefer portgroups to ports.
* Prefer ports with PXE enabled.

The change is backwards compatible, as the old behaviour is maintained
when ports have a physical_network field which is None.

Change-Id: I3d13bfacfb5578f570791e3c06e769a9a0140a4c
Partial-Bug: #1666009