Updated openstack/openstack Project: openstack/ceilometer  0972df767d6bb7922f2aaf3f23514e7345f1c19b
Limit the default numbers of metrics polled

Enabling everything to be polled and transformed by default is quite heavy and
can lead to a massive amount of unused metrics. Limit the default to a few
metrics known to be actually useful to common deployers.

We keep all pollsters enabled on unit tests and devstack.

Change-Id: Ifb5aa684e575b7904726276b7205824f2dcaa7ce