Updated openstack/openstack Project: openstack/nova  b9fec17974dcc05bd7dc9d0b43573da3d0fd72d9
Fix error message when support matrix entry is missing a driver

This was noticed in change If10cffd0dc4c9879f6754ce39bee5fae1d04f474
which was missing the powervm driver target for the extend-volume entry.

Before this change, the error message was:

  'libvirt-vz-ct' missing in '[operation.extend-volume]' section

This was really confusing because that driver is in the change. What
was missing was powervm, but because the error message is using the
wrong key that was not showing up.

Change-Id: I2e7ea49d5ba42cc633796222af47c1d4cd59f96b
Closes-Bug: #1702542