Updated openstack/openstack Project: openstack/nova  3628aaf6db66e7fa64f5f1c299b161d10745cd1d
Pre-load instance.device_metadata in InstanceMetadata

Change Ie7d97ce5c62c8fb9da5822590a64210521f8ae7a orphans
the Instance object so that we can pickle it. However,
this means we can't lazy-load any attributes that we need
when building up metadata, like when building the config
drive during server create. We need to pre-load the
device_metadata field so it's available when building
the config drive.

This isn't a problem for the libvirt driver since it
loads device_metadata before building the config drive,
but the hyperv driver doesn't do that, so the change
above breaks hyperv when there are device tags without
this fix.

Change-Id: I08b905d2734ff9d484b373369f36d48c4d056fd8
Closes-Bug: #1702150