Updated openstack/openstack Project: openstack-infra/project-config  ff6547920ad330506563872428287717acd76515
jenkins: jobs: ansible-role-jobs: Check if yum-config-manager is present

Fedora has /usr/bin/yum present but lacks yum-config-manager and as such
enabling the EPEL repository fails. Fedora does not need the EPEL
repository anyway but it's best to check if yum-config-manager is
present before we try to use it to manage the repositories.

Fixes the following problem on ansible-hardening Fedora CI tests

2017-07-11 08:41:33.826710 | [Zuul] Task exit code: 0
2017-07-11 08:41:40.250338 | + [[ -e /usr/bin/yum ]]
2017-07-11 08:41:40.250583 | + sudo yum-config-manager --enable epel
2017-07-11 08:41:40.256168 | sudo: yum-config-manager: command not found

Change-Id: I2b7b7c360197a7526e6cedc9f6c4ae5a391d9f6d