Updated openstack/openstack Project: openstack/governance  7380336669c8d165c94f96c57619cebda3fbe291
Add goal for policy and docs in code

This community goal describes the work for moving default policies
into code and documenting them. It also describes the benefits for
both operators and developers as well as some of the things we can do
with policy in the future to improve experiences with policy and RBAC
across OpenStack.

Change-Id: I424a684df9aabb9e5ed4e9ea1c0205a871c17a84