Updated openstack/openstack Project: openstack/nova  bf7203ac894d13737fa2ca18b35e9a72a49bc206
doc: Use consistent author, section for man pages

There's a lot of duplication here. Given that 'conf.py' is Python, we
can simplify this greatly and focus on the real content differences.

The description text is the same for all options, but this isn't
centralized as it shouldn't be the same (this field should contain a
description of each individual tool).

Change-Id: I1a2a55ac7f041387597ef93be62411fe1673f8c1