Updated openstack/openstack Project: openstack-infra/project-config  cdeb4da87a8349662df586421c1b4e63eadd324e
Update ACLs for stable/pike in prep for release

For cycle-with-milestones deliverables, define a specific
ACL for stable/pike pointing to the $project-release-branch
group. Pre-release, that group will include Release Managers
and $project-release (to give control to release teams).
Post-release that group will contain $project-stable-maint,
to give control to stable teams).

Change-Id: I9fa69e035d3ba50f8171b510bc9da753d48486bb