Updated openstack/openstack Project: openstack/tripleo-heat-templates  c76feac1a9d80c545330085c9928b6ff15f58096
Add DeployedServerEnvironmentOutput

Add a new output, DeployedServerEnvionmentOutput, that can be used as
the contents of an environment file to input into a services only stack
when using split-stack. The parameter simplifies the manual steps needed
to deploy split-stack.

By default, the resource that generates the output is mapped to
OS::Heat::None.

implements blueprint split-stack-default
Change-Id: I6004cd3f56778f078a69a20e93a0eba0c574b3db