Updated openstack/openstack Project: openstack/tripleo-heat-templates  0b8d09b249aa5c5daf827f59ec588d3afb413c44
Fix ironic-pxe startup issues

This patch updates the ironic-api docker service so that it
generates its config files in a unique config root. This
ensures that it doesn't have config files in the httpd
conf.d directory for both the API and PXE services thus
causing the API container to attempt to launch both of them.

This functionally fixes the Ironic API and PXE services with
docker so they once again can bring up an overcloud.

Change-Id: I537cd6a3337bf776ca38a279b7c130b6429eea04
Closes-bug: #1702799