Updated openstack/openstack Project: openstack/keystone  b3fc06c9ef6a0a3d317f414962bab3d309c2e6c0
Move import down to correct group

The "from keystone.common import utils" import was located
in the third-party group, which is the incorrect import group.

This change moves it from the group of third part imports to the
correct local import group to be more consistent with the rest
of keystone and to follow the pep8 guidelines.

Change-Id: I1d7124e67cd5a522eeafceacf05ca4ffac9d8b68