Updated openstack/openstack Project: openstack/oslo.log  3fd45abc6705e0d5a08c165061e5dd7f47d70c09
Add log.get_loggers method

In the Cinder project restarting the volume service is problematic
because it's not only in the control path but also in the data path of
some operations, so restarting it to change log levels is really
disruptive because it takes a long time to be done properly and it does
not support Active-Active configurations yet.

For this reason there's a specific API to change the log levels
dynamically based on a given prefix [1], but this feature is accessing
directly the `_loggers` dictionary which is not ideal.

This patch adds a simple method called `get_loggers` that will return a
copy of the internal `_loggers` dictionary so that the Cinder feature
can be changed to use this instead.

[1] https://review.openstack.org/445885

Change-Id: Ife33a5dd044a4565ad6a4d559ef9f0c108f260a3