Updated openstack/openstack Project: openstack/cinder  b398834a63a4f64e716cc373347a072907738400
Migrate Cinder CLI Quota docs

Another riveting chapter in the Cinder docs migration!

This patch brings over the CLI Quota related documentation
from openstack-manuals. It includes how to set quotas
as well as the Cinder specific documentation.

As with the other patches, this is just intended to migrate
the content so it is once again accessible. Improvements/changes
to the documentation can come with follow-on patches.

Change-Id: Id5a41b0ebb59ec8db9a80eb1effb873da65140f6