Updated openstack/openstack Project: openstack/tripleo-heat-templates  fb5eb77331006fe34991189b1f84c3ebb6c00132
Remove hardcoded enable_load_balancer from Controller role

This is associated with the haproxy service, so set the hieradata there
instead.  This is needed so we can render the controller role template
via j2, and also if anyone ever wants to run haproxy on some role other
then the Controller.

Change-Id: I82b992afe42f6da7788f6efca2366863c3bf68f7
Partially-Implements: blueprint composable-networks